+++ packages/create-flex-plugin/templates/js/src/components/CustomTaskList/CustomTaskList.Container.js
@@ -1,8 +1,8 @@
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 
-import { Actions } from '../states/CustomTaskListState';
-import CustomTaskListComponent from '../components/CustomTaskListComponent';
+import { Actions } from '../../states/CustomTaskListState';
+import CustomTaskList from './CustomTaskList';
 
 const mapStateToProps = (state) => ({
     isOpen: state.{{pluginNamespace}}.customTaskList.isOpen,
@@ -12,4 +12,4 @@ const mapDispatchToProps = (dispatch) => ({
   dismissBar: bindActionCreators(Actions.dismissBar, dispatch),
 });
 
-export default connect(mapStateToProps, mapDispatchToProps)(CustomTaskListComponent);
+export default connect(mapStateToProps, mapDispatchToProps)(CustomTaskList);
